refactor: use a normal ArrayList for puzzles
There doesn't seem to be much benefit to using MultiArrayList here: the number of puzzles in a puzzle set is expected to be relatively low, and we are usually focused on a single puzzle rather than on a single element across multiple puzzles. For this use-case, a normal ArrayList is a better choice and easier to use.